requirements.txt
ccxt==4.2.25
pandas==2.3.1
numpy==1.24.3
streamlit==1.29.0
plotly==5.17.0
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.7.0
aiohttp==3.9.1